refined interaction, separated content from component
